% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replr.R
\name{replr}
\alias{replr}
\title{An R REPL}
\usage{
replr(prompt = ">>>> ", continue = "++++ ")
}
\arguments{
\item{prompt}{A string. The text prompt.}

\item{continue}{A string. The text prompt when from the second line onward if
a multiline expression is passed.}
}
\description{
A simple - not yet fully functional - R REPL.
}
