% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{server_status}
\alias{server_status}
\title{Query the status of a running replr server}
\usage{
server_status(port = 8080)
}
\arguments{
\item{port}{Port number that the server is listening on.}
}
\value{
A list parsed from the JSON status response.
}
\description{
Retrieves basic status information from the server such as uptime and
version details.
}
